Help channels: move time functions to channel module
